chore: refactoring of tests, grouping and moving tests info separate files.
This is to not have one very long file with tests. No tests have been added or removed in this commit.
